new
new

+++ app/Http/Controllers/Api/JobseekerController.php
@@ -97,7 +97,7 @@ class JobseekerController extends Controller
             $request->certification_year ? $user->certification_year = implode(',', $request->certification_year) : '';
             $request->certification_description ? $user->certification_description = implode(',', $request->certification_description) : '';
             if ($request->hasfile('image')) {
-                if (!empty($user->image) && $user->image != "assets/dist/img/profilepic.png") {
+                if (!empty($user->image) && ($user->image != "assets/dist/img/profilepic.png")) {
                     $image_path = $user->image;
                     unlink($image_path);
                 }
@@ -109,7 +109,7 @@ class JobseekerController extends Controller
             }
 
             if ($request->hasfile('video')) {
-                if (!empty($user->video)) {
+                if (!empty($user->video) && ($user->video != "profile_videos/1630307178profile_video.mp4")) {
                     $video_path = $user->video;
                     unlink($video_path);
                 }
